refactor(nextwiki-infra): 重命名实体类并迁移数据库表名
- 将多个实体类由 XxxDO 重命名为 Xxx，去除 DO 后缀
- 更新对应 Mapper 接口、XML 和 Repository 实现中的类型引用
- 修改 SQL 语句中使用的数据表名称，添加 ydsz_ 前缀符合新的命名规范
- 修改代码中 FileVersionDomainService 替换为 FileVersionmainService
- 调整 application.yml 中 warmup-classes 配置，去除 DO 后缀
- 修改 AuthPolicyService 中 sessionTimeoutSeconds 类型转换为 int 类型
- 删除 ydsz-userinfo 模块下 CompanyDO、DepartmentDO、CompanyDeptDO、AuthPolicyDO 等实体类
- 更新 FileComment、FileAcl 等 Mapper 和 Repository 相关引用，去除 DO 后缀
- 将 ContentExtractionApplicationService 中 DB 降级搜索索引表名改为 ydsz_nw_search_index

+++ ydsz-userinfo/ydsz-userinfo-server/src/main/java/com/njydsz/userinfo/server/service/AuthPolicyService.java
@@ -144,7 +144,7 @@ public class AuthPolicyService {
     vo.setCaptchaEnabled(true);
     vo.setAllowedIdentityProviders("LOCAL");
     vo.setMaxSessionsPerUser(DEFAULT_MAX_SESSIONS_PER_USER);
-    vo.setSessionTimeoutSeconds(DEFAULT_SESSION_TIMEOUT_SECONDS);
+    vo.setSessionTimeoutSeconds((int) DEFAULT_SESSION_TIMEOUT_SECONDS);
     vo.setRemark("硬编码兜底策略");
     return vo;
   }
